Chore: Migrate updateMessages to Typescript (#3715)
