+++ app/definitions/IMessage.ts
@@ -54,7 +54,7 @@ export interface ILastMessage {
 }
 
 export interface IMessage {
-	_id?: string;
+	_id: string;
 	msg?: string;
 	t?: SubscriptionType;
 	ts: Date;
